requirement.txt
Flask==2.1.3
torch==1.10.0
torchvision==0.11.1
tensorflow==2.7.0
streamlit==1.12.2
Pillow==9.2.0
protobuf==3.20.1
opencv-python-headless==4.6.0.66
imutils==0.5.4
matplotlib==3.5.2
PyYAML==6.0
tqdm==4.64.0
seaborn==0.11.2
Flask-PyMongo
bcrypt